libxl: handle missing switch enum cases
Cast away enum type for libxl scheduler constants since we don't want to
cover all of them and don't want build to break when new ones are added.

+++ src/libxl/libxl_driver.c
@@ -4497,7 +4497,7 @@ libxlDomainGetSchedulerType(virDomainPtr dom, int *nparams)
 
     if (nparams)
         *nparams = 0;
-    switch (sched_id) {
+    switch ((int)sched_id) {
     case LIBXL_SCHEDULER_SEDF:
         name = "sedf";
         break;
